Renamed TrendingItemView to RepositoryItemView. Moved it to the repository package.
